refactor: group the memory storage under src/memory/
The seven modules behind "a memory" — Lance mechanics, the two object-store decorators,
the Hub client and commit plumbing, the dataset card, the writer lock — were scattered
across the top of src/ under names (hub, dataset, card) that said nothing about which
layer they belong to. They move under src/memory/, whose root records the split the
consolidation will act on: mechanics (dataset, fetch_store, capture_store), transport
(hf_dataset), domain (hub).

Pure move: module declarations, intra-layer references switched to `use super::…`, and
grouped `use crate::{…}` lines split so the memory modules come from one path.
