+++ scripts/backfill_remote_repo.rs
@@ -12,8 +12,8 @@
 use anyhow::{bail, Context, Result};
 use arrow_array::{RecordBatch, StringArray};
 use arrow_schema::{DataType, Field, Schema};
+use funes::memory::{dataset, hf_dataset, hub};
 use funes::traces::repo;
-use funes::{dataset, hf_dataset, hub};
 use hf_hub::HFClient;
 use lance::dataset::{BatchUDF, Dataset, NewColumnTransform};
 use std::collections::HashMap;
