+++ src/curate.rs
@@ -18,9 +18,10 @@
 //! Human- and agent-editable. A decision flipped to `exclude` later does not retract what
 //! already shipped — the remote is append-only; curation prevents, it does not undo.
 
-use crate::hub::{self, Memory};
+use crate::index;
+use crate::memory::hub::{self, Memory};
+use crate::memory::{dataset, hf_dataset};
 use crate::traces::jsonl;
-use crate::{dataset, hf_dataset, index};
 use anyhow::{bail, Context, Result};
 use arrow_array::{Int64Array, RecordBatch, RecordBatchIterator, StringArray};
 use arrow_schema::Schema;
