+++ src/lib.rs
@@ -16,18 +16,12 @@ pub mod ask;
 pub mod banner;
 #[cfg(feature = "blas")]
 pub mod blas;
-pub mod capture_store;
-pub mod card;
 pub mod chunk;
 pub mod curate;
-pub mod dataset;
-pub mod fetch_store;
-pub mod hf_dataset;
-pub mod hub;
 pub mod index;
 pub mod inference;
-pub mod lock;
 pub mod mcp;
+pub mod memory;
 pub mod push;
 pub mod recall;
 pub mod render;
