+++ src/memory/capture_store.rs
