+++ tests/push_card.rs
@@ -15,7 +15,7 @@ use std::io::Write;
 use std::process::Command;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use funes::hub::Memory;
+use funes::memory::hub::Memory;
 use funes::push::Confirm;
 use hf_hub::repository::CommitOperation;
 use hf_hub::{HFClient, HFError, HFRepository, RepoTypeDataset};
@@ -74,7 +74,7 @@ async fn card_created_refreshed_and_a_hand_card_respected() {
     // A throwaway ROOT memory: unique repo name so concurrent/repeated runs don't collide.
     let nanos = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos();
     let name = format!("funes-test-card-{}-{nanos}", std::process::id());
-    if let Err(e) = funes::hub::create_dataset_repo(OWNER, &name).await {
+    if let Err(e) = funes::memory::hub::create_dataset_repo(OWNER, &name).await {
         eprintln!("skip: cannot create a scratch repo under {OWNER}: {e}");
         return;
     }
